+++ extensions/iceworks-ui-builder/src/extension.ts
@@ -4,6 +4,7 @@ import { initExtension, registerCommand } from '@iceworks/common-service';
 import { Recorder } from '@iceworks/recorder';
 import services from './services/index';
 import i18n from './i18n';
+import { registerDebugCommand } from './debugMaterials';
 
 // eslint-disable-next-line
 const { name, version } = require('../package.json');
@@ -107,6 +108,7 @@ export function activate(context: vscode.ExtensionContext) {
       activePageCreatorWebview();
     }),
   );
+  registerDebugCommand(subscriptions);
 }
 
 export function deactivate() { }