+++ packages/project-service/src/index.ts
@@ -147,16 +147,18 @@ export async function getScaffolds(source: string) {
   return response.data.scaffolds;
 }
 
-export async function getFolderPath(): Promise<string> {
+export async function getFolderPath(openLabel = 'Open'): Promise<string | undefined> {
   const options: vscode.OpenDialogOptions = {
     canSelectFolders: true,
     canSelectFiles: false,
     canSelectMany: false,
-    openLabel: 'Open',
+    openLabel,
   };
   const selectFolderUri = await vscode.window.showOpenDialog(options);
-  const { fsPath } = selectFolderUri[0];
-  return fsPath;
+  if (selectFolderUri) {
+    const { fsPath } = selectFolderUri[0];
+    return fsPath;
+  }
 }
 
 export async function createProject(projectField: IProjectField): Promise<string> {