+++ packages/vscode-generator/bin/config.js
@@ -1,0 +1,20 @@
+function getPromptQuestion() {
+  const promptQuestion = [{
+    type: 'list',
+    name: 'projectType',
+    message: 'What\'s your project type?',
+    default: 'extension',
+    choices: [
+      {
+        name: 'VS Code Extension',
+        value: 'extension',
+      },
+    ],
+  }];
+
+  return promptQuestion;
+}
+
+module.exports = {
+  getPromptQuestion,
+};