+++ packages/vscode-generator/src/templates/extension/web/src/layouts/index.tsx
@@ -1,0 +1,66 @@
+import React, { useState } from 'react';
+import { Shell, ConfigProvider } from '@alifd/next';
+
+(function () {
+  const throttle = function (type: string, name: string, obj: Window = window) {
+    let running = false;
+
+    const func = () => {
+      if (running) {
+        return;
+      }
+
+      running = true;
+      requestAnimationFrame(() => {
+        obj.dispatchEvent(new CustomEvent(name));
+        running = false;
+      });
+    };
+
+    obj.addEventListener(type, func);
+  };
+
+  if (typeof window !== 'undefined') {
+    throttle('resize', 'optimizedResize');
+  }
+})();
+
+interface IGetDevice {
+  (width: number): 'phone' | 'tablet' | 'desktop';
+}
+export default function BasicLayout({ children }: { children: React.ReactNode }) {
+  const getDevice: IGetDevice = (width) => {
+    const isPhone = typeof navigator !== 'undefined' && navigator && navigator.userAgent.match(/phone/gi);
+
+    if (width < 680 || isPhone) {
+      return 'phone';
+    } else if (width < 1280 && width > 680) {
+      return 'tablet';
+    } else {
+      return 'desktop';
+    }
+  };
+
+  const [device, setDevice] = useState(getDevice(NaN));
+
+  if (typeof window !== 'undefined') {
+    window.addEventListener('optimizedResize', (e) => {
+      const deviceWidth = (e && e.target && (e.target as Window).innerWidth) || NaN;
+      setDevice(getDevice(deviceWidth));
+    });
+  }
+
+  return (
+    <ConfigProvider device={device}>
+      <Shell
+        style={{
+          overflow: 'auto',
+          minWidth: '350px',
+          minHeight: '100vh',
+        }}
+      >
+        <Shell.Content style={{ padding: '8px 10px' }}>{children}</Shell.Content>
+      </Shell>
+    </ConfigProvider>
+  );
+}