+++ extensions/iceworks-doctor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.1.7
+
+- chore: change extension name
+
+## 0.1.6
+
+- fix: [#604](https://github.com/ice-lab/iceworks/issues/604)
+
+## 0.1.5
+
+- feat: add `Iceworks: Scan Your Project` command.
+
 ## 0.1.3
 
 - fix: compatible with scenarios without material source